Getting rid of UINT, UINT8, UINT32, ULONG types (#670, #23)

+++ src/xrCore/_math.cpp
@@ -81,12 +81,12 @@ public:
 
 typedef struct _PROCESSOR_POWER_INFORMATION
 {
-    ULONG Number;
-    ULONG MaxMhz;
-    ULONG CurrentMhz;
-    ULONG MhzLimit;
-    ULONG MaxIdleState;
-    ULONG CurrentIdleState;
+    u32 Number;
+    u32 MaxMhz;
+    u32 CurrentMhz;
+    u32 MhzLimit;
+    u32 MaxIdleState;
+    u32 CurrentIdleState;
 } PROCESSOR_POWER_INFORMATION, *PPROCESSOR_POWER_INFORMATION;
 
 // Initialized on startup
